Extract shared services to enable cross-package testing
Refactors common service interfaces and implementations into a shared package
to improve code reuse and enable comprehensive testing across backend and
frontend components:

• Create shared/services.go with common service interfaces and implementations
• Move ContextManager, CommandRunner, RuntimeInfo, and UserInfo to shared package
• Update grep tool to accept CommandRunner parameter for dependency injection
• Refactor frontend CLI commands to use shared service implementations
• Add CommandRunner field to codeact Session for testable command execution
• Update all import statements to reference shared package implementations

This refactoring enables better unit testing by allowing mock injection of
system dependencies and eliminates code duplication between backend tools
and frontend CLI commands.
